+++ src/main/java/team_alcoholic/jumo_server/domain/auth/dto/CustomOAuth2User.java
@@ -61,8 +61,17 @@ public class CustomOAuth2User implements OAuth2User, Serializable {
         return userDTO.getProvider() + " " + userDTO.getProviderId();
     }
 
+    /**
+     * 신규 사용자인지 구분하기 위해 isNewUser를 반환하는 메서드
+     * @return
+     */
     public boolean isNewUser() { return this.isNewUser; }
 
+    /**
+     * 갱신된 회원정보를 담고 있는 UserDTO를 받아옴
+     * 회원정보 수정 시 세션 정보를 함께 수정할 때 사용
+     * @param newUserDTO
+     */
     public void updateAttributes(UserDTO newUserDTO) {
         this.userDTO = newUserDTO;
     }
