+++ src/main/java/team_alcoholic/jumo_server/global/config/SecurityConfig.java
@@ -1,7 +1,6 @@
 package team_alcoholic.jumo_server.global.config;
 
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -11,10 +10,9 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.access.channel.ChannelProcessingFilter;
-import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
-import team_alcoholic.jumo_server.domain.auth.dto.CustomOAuth2User;
 import team_alcoholic.jumo_server.domain.auth.service.OAuth2UserService;
-import org.springframework.security.web.savedrequest.RequestCache;
+import team_alcoholic.jumo_server.global.config.oauth2.CustomOAuth2SuccessHandler;
+import team_alcoholic.jumo_server.global.config.oauth2.GetRedirectUrlFilter;
 
 
 @Configuration
@@ -24,6 +22,7 @@ public class SecurityConfig {
 
     private final OAuth2UserService oAuth2UserService;
     private final CustomAuthenticationEntryPoint customAuthenticationEntryPoint;
+    private final CustomOAuth2SuccessHandler customOAuth2SuccessHandler;
 
     @Value("${service.url}")
     private String serviceUrl;
@@ -32,43 +31,24 @@ public class SecurityConfig {
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
 
         http
-                .csrf(AbstractHttpConfigurer::disable)
-                .formLogin(AbstractHttpConfigurer::disable)
-                .httpBasic(AbstractHttpConfigurer::disable)
-                .addFilterBefore(new GetRedirectUrlFilter(), ChannelProcessingFilter.class) // 필터 체인의 맨 앞에 필터 추가
-                .oauth2Login(oauth2 -> oauth2
-                        .userInfoEndpoint(userInfoEndpoint -> userInfoEndpoint.userService(oAuth2UserService))
-                        .successHandler((request, response, authentication) -> {
-                            CustomOAuth2User customOAuth2User = (CustomOAuth2User) authentication.getPrincipal();
-                            if (customOAuth2User.isNewUser()) {
-                                response.sendRedirect(serviceUrl + "/join");
-                            } else {
-                                HttpSession session = request.getSession();
-                                String redirectUrl = (String) session.getAttribute("redirectUrl");
-                                if (redirectUrl == null || redirectUrl.isEmpty()) {
-                                    redirectUrl = serviceUrl;
-                                }
-                                response.sendRedirect(redirectUrl);
-                            }
-                        }))
-                .exceptionHandling(exceptionHandling -> exceptionHandling
-                        .authenticationEntryPoint(customAuthenticationEntryPoint)
-                )
-                .logout(logout -> logout
-                        .logoutUrl("/logout")
-                        .invalidateHttpSession(true)
-                        .deleteCookies("SESSION")
-                        .logoutSuccessHandler((request, response, authentication) -> {
-                            response.setStatus(HttpServletResponse.SC_OK);
-                        })
-                );
-
-//        http
-//                .authorizeHttpRequests(auth -> auth
-//                        .requestMatchers(String.valueOf(HttpMethod.POST), "/tasting-notes").authenticated()
-//                        .anyRequest().permitAll()
-//                );
-
+            .csrf(AbstractHttpConfigurer::disable)
+            .formLogin(AbstractHttpConfigurer::disable)
+            .httpBasic(AbstractHttpConfigurer::disable)
+            .addFilterBefore(new GetRedirectUrlFilter(), ChannelProcessingFilter.class) // 필터 체인의 맨 앞에 필터 추가
+            .oauth2Login(oauth2 -> oauth2
+                .userInfoEndpoint(userInfoEndpoint -> userInfoEndpoint.userService(oAuth2UserService))
+                .successHandler(customOAuth2SuccessHandler))
+            .exceptionHandling(exceptionHandling -> exceptionHandling
+                .authenticationEntryPoint(customAuthenticationEntryPoint)
+            )
+            .logout(logout -> logout
+                .logoutUrl("/logout")
+                .invalidateHttpSession(true)
+                .deleteCookies("SESSION")
+                .logoutSuccessHandler((request, response, authentication) -> {
+                    response.setStatus(HttpServletResponse.SC_OK);
+                })
+            );
         return http.build();
     }
